updated defaults  for dtypes

+++ qasr/misc/misc_utils.py
@@ -22,7 +22,7 @@ def parse_args():
                          help='Max num of tokens to generate')
     parser.add_argument('--max_think_tokens', type=int, default=256,
                          help='Max num of tokens to generate')
-    parser.add_argument('--model_dtype', type=str, default='float16',
+    parser.add_argument('--model_dtype', type=str, default='bfloat16',
                         choices=['auto', 'bfloat16', 'float16', 'float32'])
 
     parser.add_argument('--quant_config', type=str, default=None,
@@ -39,7 +39,7 @@ def parse_args():
     parser.add_argument('--dataset_path', type=str, default='hf-audio/esb-datasets-test-only-sorted')
     parser.add_argument('--dataset', type=str, default='voxpopuli')
     parser.add_argument('--split', type=str, default='test')
-    parser.add_argument('--data_dtype', type=str, default='float16',
+    parser.add_argument('--data_dtype', type=str, default='bfloat16',
                          choices=['bfloat16', 'float16', 'float32'])
 
     parser.add_argument('--batch_size', type=int, default=64)
